refactoredd folder structure
