+++ src/pages/teacher/dashboard/create-course/basic details/Crop.tsx
