+++ src/pages/teacher/dashboard/create-course/basic details/Thumbnail.tsx
@@ -2,7 +2,7 @@ import { useUpdateThumnailMutation } from "@/redux/features/teacher/courseCreati
 import { Pencil, Upload } from "lucide-react";
 import React, { useRef } from "react";
 import toast from "react-hot-toast";
-import { ICourse } from "./CourseDraft";
+import { ICourse } from "../CourseDraft";
 import ImgCrop from "./Crop";
 import {
     Dialog,
