+++ src/pages/teacher/dashboard/create-course/basic details/canvasPreview.ts
