+++ src/pages/teacher/dashboard/create-course/course-assets/CourseAssetsOutlet.tsx
